Use redis lists for doc editor

See merge request upchieve/subway!939

+++ server/services/SessionService.ts
@@ -50,6 +50,7 @@ import {
   beginRegularNotifications,
 } from './TwilioService'
 import * as WhiteboardService from './WhiteboardService'
+import { LockError } from 'redlock'
 
 export async function reviewSession(data: unknown) {
   const { sessionId, reviewed, toReview } = sessionUtils.asReviewSessionData(
@@ -318,10 +319,30 @@ export async function processFirstSessionCongratsEmail(
   }
 }
 
-export async function storeAndDeleteQuillDoc(sessionId: Types.ObjectId) {
-  const quillDoc = await QuillDocService.getDoc(sessionId)
-  await SessionRepo.updateSessionQuillDoc(sessionId, JSON.stringify(quillDoc))
-  await QuillDocService.deleteDoc(sessionId)
+export async function storeAndDeleteQuillDoc(
+  sessionId: Types.ObjectId,
+  retries: number = 0
+): Promise<void> {
+  let quillState: QuillDocService.QuillCacheState | undefined
+  try {
+    quillState = await QuillDocService.lockAndGetDocCacheState(sessionId)
+  } catch (error) {
+    if (error instanceof LockError && retries < 10)
+      return storeAndDeleteQuillDoc(sessionId, retries + 1)
+    else
+      logger.error(
+        `Failed to update and get document in the cache for session ${sessionId} - ${error}`
+      )
+    return
+  }
+
+  if (quillState) {
+    await SessionRepo.updateSessionQuillDoc(
+      sessionId,
+      JSON.stringify(quillState.doc)
+    )
+    await QuillDocService.deleteDoc(sessionId)
+  }
 }
 
 export async function storeAndDeleteWhiteboardDoc(sessionId: Types.ObjectId) {
@@ -520,7 +541,6 @@ export async function adminSessionView(data: unknown) {
   )
   const feedback = await getFeedbackV2BySessionId(session._id)
   const bucket: keyof typeof config.awsS3 = 'sessionPhotoBucket'
-  let s3Keys
   const sessionPhotos = await AwsService.getObjects(bucket, session.photos)
 
   return {
